+++ resources/views/account/addresses.blade.php
@@ -1,0 +1,3 @@
+<x-layouts::storefront :title="__('My Addresses')">
+    <livewire:storefront.address-book />
+</x-layouts::storefront>